refactor: complete session 3 file deduplication
