+++ src/core/teleportLogic.ts
@@ -1,7 +1,7 @@
 import { errorLog } from '@core/logger.js';
 import { playSound } from '@core/utils/sound.js';
 import { getCountdownColor } from '@core/utils/ui.js';
-import { setActionBarOverride } from '@features/sidebar/sidebarManager.js';
+import { setActionBarOverride } from '@features/sidebar/manager.js';
 import { isDefined, isNumber } from '@lib/guards.js';
 import { Vector3Utils } from '@minecraft/math';
 import * as mc from '@minecraft/server';
