+++ src/core/ui/panels/index.ts
@@ -2,13 +2,13 @@ import { BountyPanelHandler } from '@features/economy/ui/bountyPanel.js';
 import { EconomyPanelHandler } from '@features/economy/ui/panel.js';
 import { WorldProtectionPanelHandler } from '@features/essentials/ui/worldProtectionPanel.js';
 import { KitPanelHandler } from '@features/kit/ui/panel.js';
-import { ModerationPanelHandler } from '@features/moderation/ui/moderationPanel.js';
+import { ModerationPanelHandler } from '@features/moderation/ui/panel.js';
 import { XrayPanelHandler } from '@features/moderation/ui/xrayPanel.js';
-import { ShopAdminPanelHandler } from '@features/shop/ui/shopAdminPanel.js';
-import { ShopUserPanelHandler } from '@features/shop/ui/shopUserPanel.js';
+import { ShopAdminPanelHandler } from '@features/shop/ui/adminPanel.js';
+import { ShopUserPanelHandler } from '@features/shop/ui/userPanel.js';
 import { FriendPanelHandler } from '@features/social/ui/friendPanel.js';
-import { TeamPanelHandler } from '@features/team/ui/teamPanel.js';
-import { TeleportPanelHandler } from '@features/teleport/ui/teleportPanel.js';
+import { TeamPanelHandler } from '@features/team/ui/panel.js';
+import { TeleportPanelHandler } from '@features/teleport/ui/panel.js';
 import { panelRouter } from '@ui/PanelRouter.js';
 import { SidebarPanelHandler } from '@ui/panels/sidebarPanel.js';
 import { AdminPanelHandler } from './adminPanel.js';
