+++ src/features/shop/manager.ts
