+++ src/features/teleport/commands/tp.ts
@@ -3,7 +3,7 @@ import * as mc from '@minecraft/server';
 import { CustomCommand } from '@commands/commandManager.js';
 import { sendMessage } from '@core/messaging.js';
 import { findPlayerByName } from '@core/playerCache.js';
-import { saveLastLocation } from '@features/teleport/teleportUtils.js';
+import { saveLastLocation } from '@features/teleport/utils.js';
 import { isDefined, isNonEmptyString, isNumber } from '@lib/guards.js';
 
 const command: CustomCommand = {
