+++ src/features/teleport/index.ts
@@ -1,4 +1,4 @@
-import { TeleportPanelHandler } from '@features/teleport/ui/teleportPanel.js';
+import { TeleportPanelHandler } from '@features/teleport/ui/panel.js';
 import { panelRouter } from '@ui/PanelRouter.js';
 
 export function initialize() {
