+++ src/features/vote/manager.ts
