+++ src/features/vote/ui/panel.ts
@@ -4,7 +4,7 @@ import { ActionFormData, ActionFormResponse, ModalFormData, ModalFormResponse } 
 import { getConfig } from '@core/configManager.js';
 import { getPlayerRank } from '@core/rankManager.js';
 import { uiWait } from '@core/utils.js';
-import { castVote, createVote, endVote, getActiveVote, getLastVote } from '@features/vote/voteManager.js';
+import { castVote, createVote, endVote, getActiveVote, getLastVote } from '@features/vote/manager.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 
 export async function showVoteMenu(player: mc.Player) {
